+++ src/main/java/duke/Duke.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.command.Command;
+import duke.exception.DukeException;
+
 public class Duke {
 
     private final Storage storage;
