+++ src/main/java/duke/Parser.java
@@ -1,3 +1,20 @@
+package duke;
+
+import duke.command.Command;
+import duke.command.Action;
+import duke.command.List;
+import duke.command.Add;
+import duke.command.End;
+import duke.exception.DukeException;
+import duke.exception.MissingDescriptionException;
+import duke.exception.UnknownCommandException;
+import duke.exception.InvalidFormatException;
+import duke.exception.InvalidDateTimeException;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.ToDo;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
