+++ src/main/java/duke/command/End.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class End extends Command {
 
     @Override
