+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class DukeException extends Exception {
+    DukeException(String errorMsg) {
+        super(errorMsg);
+    }
+}
+
