+++ src/main/java/duke/exception/InvalidFormatException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidFormatException extends DukeException {
+    public InvalidFormatException(String format) {
+        super("Invalid format! Try " + format);
+    }
+}
