+++ src/main/java/duke/exception/InvalidIntegerException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidIntegerException extends DukeException {
+    public InvalidIntegerException() {
+        super("Enter valid integer!");
+    }
+}
