+++ src/main/java/duke/exception/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidTaskNumberException extends DukeException {
+    public InvalidTaskNumberException() {
+        super("Not a valid task number!");
+    }
+}
