+++ src/main/java/duke/exception/LoadingError.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class LoadingError extends DukeException {
+    public LoadingError(String message) {
+        super("Loading Error: " + message);
+    }
+}
