+++ src/main/java/duke/exception/MissingDescriptionException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class MissingDescriptionException extends DukeException {
+    public MissingDescriptionException() {
+        super("A task requires a description following it!");
+    }
+}
