+++ src/main/java/duke/exception/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class UnknownCommandException extends DukeException {
+    public UnknownCommandException() {
+        super("I'm not too sure what you mean :(");
+    }
+}
