+++ src/main/java/duke/tasks/Event.java
@@ -1,20 +1,21 @@
+package duke.tasks;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
 
     protected LocalDate date;
     protected LocalTime time;
 
-    public Deadline(String description, LocalDate date) {
+    public Event(String description, LocalDate date) {
         super(description);
         this.date = date;
         this.time = null;
-
     }
 
-    public Deadline(String description, LocalDate date, LocalTime time) {
+    public Event(String description, LocalDate date, LocalTime time) {
         super(description);
         this.date = date;
         this.time = time;
@@ -24,10 +25,10 @@ public class Deadline extends Task {
     public String toString() {
         String dateString = this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
         if (this.time == null) {
-            return "[D]" + super.toString() + " (by: " + dateString + ")";
+            return "[E]" + super.toString() + " (at: " + dateString + ")";
         } else {
             String timeString = this.time.format(DateTimeFormatter.ofPattern("hh:mm a"));
-            return "[D]" + super.toString() + " (by: " + dateString + " " + timeString + ")";
+            return "[E]" + super.toString() + " (at: " + dateString + " " + timeString + ")";
         }
     }
 
@@ -35,10 +36,10 @@ public class Deadline extends Task {
     public String toSaveString() {
         String dateString = this.date.format(DateTimeFormatter.ISO_LOCAL_DATE);
         if (this.time == null) {
-            return "| D | " + super.toSaveString() + " | " + dateString;
+            return "| E | " + super.toSaveString() + " | " + dateString;
         } else {
             String timeString = this.time.format(DateTimeFormatter.ofPattern("HH:mm"));
-            return "| D | " + super.toSaveString() + " | " + dateString + " | " + timeString;
+            return "| E | " + super.toSaveString() + " | " + dateString + " | " + timeString;
         }
     }
 }
