+++ text-ui-test/input.txt
@@ -7,6 +7,8 @@ todo
 todo
 event conference /at
 deadline submission friday
+deadline wah /by 2021-08-22
+event hah /at 2021-08-29 11:30
 Hi
 list
 done 0
@@ -17,6 +19,9 @@ done 4
 delete him
 delete 1
 delete 100
-delete
+delete 1
+delete 1
+delete 1
+delete 1
 list
 bye
